Update to use a static file path
It appears that Visual Studio falls back to using the default settings
instead of the one located by the searcher. This tends to be a common
problem with Visual Studio where the experience in the IDE partially
mimics the build, but occasionally has differences.

The fix is to detect and generate the file in the intermediate output path
then always reference this file which should fix the VS weirdness